+++ src/main/java/team/incude/gsmc/v2/domain/auth/application/usecase/service/RefreshService.java
@@ -8,7 +8,7 @@ import team.incude.gsmc.v2.domain.auth.exception.RefreshTokenInvalidException;
 import team.incude.gsmc.v2.domain.auth.presentation.data.response.AuthTokenResponse;
 import team.incude.gsmc.v2.domain.member.application.port.MemberPersistencePort;
 import team.incude.gsmc.v2.domain.member.domain.Member;
-import team.incude.gsmc.v2.global.security.jwt.dto.TokenDto;
+import team.incude.gsmc.v2.global.security.jwt.data.TokenDto;
 
 /**
  * 리프레시 토큰을 검증하고 새로운 인증 토큰을 발급하는 유스케이스 구현 클래스입니다.
